Download anor databases

Download anor databases

download.database(
  download.name = NULL,
  database.dir = tempdir(),
  version = c(),
  buildver = "hg19",
  download.cfg = system.file("extdata", "config/db/db_annovar.toml", package =
    "BioInstaller"),
  show.all.versions = FALSE,
  show.all.names = FALSE,
  show.all.buildvers = FALSE,
  verbose = FALSE,
  ...
)

Arguments

download.name
Download name, default is NULL and get value from database.cfg dependence_db

database.dir
Destination directory to store databases

version
Version of download database, supported version can be get by download.database('db_annovar_avsnp', show.all.versions = TRUE)

buildver
Genome version, e.g hg19, hg38, mm10

download.cfg
Configuration file for download, most of anor database resources can be found in system.file('extdata', 'config/db/db_annovar.toml', package = 'BioInstaller')

show.all.versions
Logical indicating wheather show all avaliable versions

show.all.names
Logical indicating wheather show all avaliable name

show.all.buildvers
Logical indicating wheather show all avaliable buildver

verbose
Logical indicating wheather print the extra log infomation

...
Other parameters pass to install.bioinfo

Examples

download.database('db_annovar_avsnp', database.dir = sprintf('%s/databases/', tempdir()),
show.all.versions = TRUE)

#> [1] "avsnp150" "avsnp147" "avsnp144" "avsnp142" "avsnp138"