Build anor database in mysql

Build anor database in mysql

mysql.build(
  filename = "",
  mysql.connect.params = list(host = "", dbname = "", table.name = "", user = "",
    password = ""),
  dat = data.table(),
  fread.params = list(),
  new.colnames = NULL,
  verbose = FALSE,
  ...
)

Arguments

filename
Path of raw data, will be read by fread

mysql.connect.params
Mysql parameters, [host, dbname, table.name, user, password etc.]

dat
Object of data.table, as the input data to build database

fread.params
Other parameters be used in fread

new.colnames
New colnames of table, default is to retain the original

verbose
Ligical indicating wheather show the log message

...
Other parameters be used in dbWriteTable

Examples

test.dat <- system.file('extdata', 'demo/sqlite.dat.txt', package = 'anor')
if (FALSE) {
mysql.build(test.dat, list(host = 'host', dbname = 'db',
table.name = 'table', user = 'user', password = 'password'))
}