R function to run vcfanno

R function to run vcfanno

vcfanno(
  vcfanno = Sys.which(c("vcfanno", "vcfanno_osx", "vcfanno_linux64")),
  vcfanno_database_cfg = system.file("extdata", "demo/vcfanno_demo/conf.toml", package
    = "anor"),
  base_path = "",
  lua = "",
  ends = FALSE,
  input.file = "input.vcf",
  out = "output.vcf",
  thread = 2,
  permissive_overlap = FALSE,
  debug = FALSE
)

Arguments

vcfanno
Executable file of vcfanno (Download from https://github.com/brentp/vcfanno/releases)

vcfanno_database_cfg
vcfanno required database configuration file (Not the anor database.cfg)

base_path
Optional base_path to prepend to annotation files in the config

lua
Optional path to a file containing custom javascript functions to be used as ops

ends
Annotate the start and end as well as the interval itself.

input.file
Input file path (VCF only)

out
Output file path

thread
number of processes to use. (default 2)

permissive_overlap
annotate with an overlapping variant even it doesn't share the same ref and alt alleles. Default is to require exact match between variants.

debug
If set TRUE, only print the command

Examples

vcfanno(debug = TRUE)

#> vcfanno_linux64 -p 2 /tmp/Rtmp7z5Wx3/temp_libpath5a6a96c49a08/anor/extdata/demo/vcfanno_demo/conf.toml input.vcf > output.vcf

#> [1] "vcfanno_linux64 -p 2 /tmp/Rtmp7z5Wx3/temp_libpath5a6a96c49a08/anor/extdata/demo/vcfanno_demo/conf.toml input.vcf > output.vcf"